feat(Lotto): 제공된 Lotto 클래스에 맞게 구현
- LottoSheet 삭제
- Lotto에 맞게 구현
- Lotto 패키지 이동
